cleaning up the front-end a little bit
