+++ client/src/pages/FindService.js
@@ -1,23 +1,10 @@
 import { Link } from 'react-router-dom';
 import Auth from '../utils/auth';
-import "./FindService.css";
+import "./css/FindService.css";
 
 const FindService = () =>{
     return(
         <main className="base-grid home-columns">
-            <nav className="full-width nav-columns distribute-even fit">
-                <Link to="/profile">
-                <button className="btn">Profile</button>
-                </Link>
-                <Link to="/find-service">
-                <button className="btn">Find Service</button>
-                </Link>
-                <Link to="/offer-service">
-                <button className="btn">Offer Service</button>
-                </Link>
-                <button className="btn">Language</button>
-                <button onClick={Auth.logout}className="btn">Logout</button>
-            </nav>
             <div className="full-width">
                 <h3 className="center">Find a service</h3>
             </div>
