+++ client/src/pages/OfferService.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useMutation } from '@apollo/client'
 import { Link } from 'react-router-dom';
 import Auth from '../utils/auth';
-import "./OfferService.css";
+import "./css/OfferService.css";
 import { useQuery } from '@apollo/client';
 import {ADD_SERVICEPOST} from '../utils/mutations';
 import { QUERY_SERVICES, QUERY_ME } from '../utils/queries'
@@ -88,19 +88,6 @@ const OfferService = () =>{
 
     return(
     <main className="base-grid home-columns">
-        <nav className="full-width nav-columns distribute-even fit">
-            <Link to="/profile">
-            <button className="btn">Profile</button>
-            </Link>
-            <Link to="/find-service">
-            <button className="btn">Find Service</button>
-            </Link>
-            <Link to="/offer-service">
-            <button className="btn">Offer Service</button>
-            </Link>
-            <button className="btn">Language</button>
-            <button onClick={Auth.logout}className="btn">Logout</button>
-        </nav>
         <div className="full-width">
                 <h3 className="center">Offer a service</h3>
             </div>
